:rotating_light: update and clean up tests
